move TodoContainer outside of the components folder
