+++ src/App.js
@@ -1,4 +1,4 @@
-import TodoContainer from "./components/TodoContainer";
+import TodoContainer from "./TodoContainer";
 
 export default function App() {
   return (
